+++ okta/spCertificate.go
@@ -18,9 +18,6 @@
 
 package okta
 
-type InlineHookChannelConfig struct {
-	AuthScheme *InlineHookChannelConfigAuthScheme `json:"authScheme,omitempty"`
-	Headers    []*InlineHookChannelConfigHeaders  `json:"headers,omitempty"`
-	Method     string                             `json:"method,omitempty"`
-	Uri        string                             `json:"uri,omitempty"`
+type SpCertificate struct {
+	X5c []string `json:"x5c,omitempty"`
 }