+++ tests/integration/group_test.go
@@ -22,9 +22,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/okta/okta-sdk-golang/v2/okta/query"
-
 	"github.com/okta/okta-sdk-golang/v2/okta"
+	"github.com/okta/okta-sdk-golang/v2/okta/query"
 	"github.com/okta/okta-sdk-golang/v2/tests"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -35,7 +34,7 @@ func Test_can_get_a_group(t *testing.T) {
 	require.NoError(t, err)
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "Get Test Group",
+		Name: "SDK_TEST Get Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -65,7 +64,7 @@ func Test_can_list_groups(t *testing.T) {
 	require.NoError(t, err)
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "List Test Group",
+		Name: "SDK_TEST List Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -95,7 +94,7 @@ func Test_can_search_for_a_group(t *testing.T) {
 	require.NoError(t, err)
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "Search Test Group",
+		Name: "SDK_TEST Search Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -105,7 +104,7 @@ func Test_can_search_for_a_group(t *testing.T) {
 	assert.IsType(t, &okta.Group{}, group)
 
 	// Search the group by name with query parameter → GET /api/v1/groups?q=Search
-	groupList, _, err := client.Group.ListGroups(ctx, query.NewQueryParams(query.WithQ("Search Test Group")))
+	groupList, _, err := client.Group.ListGroups(ctx, query.NewQueryParams(query.WithQ("SDK_TEST Search Test Group")))
 	assert.Len(t, groupList, 1, "Did not find correct amount of groups")
 	require.NoError(t, err, "Listing groups should not error")
 	found := false
@@ -126,7 +125,7 @@ func Test_can_update_a_group(t *testing.T) {
 	require.NoError(t, err)
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "Update Test Group",
+		Name: "SDK_TEST Update Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -137,14 +136,14 @@ func Test_can_update_a_group(t *testing.T) {
 
 	// Update the group name and description → PUT /api/v1/groups/{{groupId}}
 	ngp := &okta.GroupProfile{
-		Name: "Updated Name",
+		Name: "SDK_TEST Updated Name",
 	}
 	client.Group.UpdateGroup(ctx, group.Id, okta.Group{Profile: ngp})
 
 	// Verify that group profile is updated by calling get on the group and verifying the profile → GET /api/v1/groups/{{groupId}}
 	updatedGroup, _, err := client.Group.GetGroup(ctx, group.Id)
 	require.NoError(t, err, "Should not error when getting updated group")
-	assert.Equal(t, "Updated Name", updatedGroup.Profile.Name, "The group was not updated")
+	assert.Equal(t, "SDK_TEST Updated Name", updatedGroup.Profile.Name, "The group was not updated")
 
 	// Delete the group → DELETE /api/v1/groups/{{groupId}}
 	_, err = client.Group.DeleteGroup(ctx, group.Id)
@@ -164,8 +163,8 @@ func Test_group_user_operations(t *testing.T) {
 	profile := okta.UserProfile{}
 	profile["firstName"] = "John"
 	profile["lastName"] = "With-Group"
-	profile["email"] = "john-with-group@example.com"
-	profile["login"] = "john-with-group@example.com"
+	profile["email"] = randomEmail()
+	profile["login"] = profile["email"]
 	u := &okta.CreateUserRequest{
 		Credentials: uc,
 		Profile:     &profile,
@@ -178,7 +177,7 @@ func Test_group_user_operations(t *testing.T) {
 
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "Group-Member API Test Group",
+		Name: "SDK_TEST Group-Member API Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -227,8 +226,8 @@ func Test_group_rule_operations(t *testing.T) {
 	profile := okta.UserProfile{}
 	profile["firstName"] = "John"
 	profile["lastName"] = "With-Group-Rule"
-	profile["email"] = "john-with-group-rule@example.com"
-	profile["login"] = "john-with-group-rule@example.com"
+	profile["email"] = randomEmail()
+	profile["login"] = profile["email"]
 	u := &okta.CreateUserRequest{
 		Credentials: uc,
 		Profile:     &profile,
@@ -241,7 +240,7 @@ func Test_group_rule_operations(t *testing.T) {
 
 	// Create a new group → POST /api/v1/groups
 	gp := &okta.GroupProfile{
-		Name: "Group-Member-Rule API Test Group",
+		Name: "SDK_TEST Group-Member-Rule API Test Group",
 	}
 	g := &okta.Group{
 		Profile: gp,
@@ -270,7 +269,7 @@ func Test_group_rule_operations(t *testing.T) {
 		Actions:    gra,
 		Conditions: grc,
 		Type:       "group_rule",
-		Name:       "Test group rule",
+		Name:       "SDK_TEST group rule",
 	}
 	groupRule, _, err := client.Group.CreateGroupRule(ctx, *gr)
 	require.NoError(t, err, "Should not error when creating a group Rule")
@@ -323,7 +322,7 @@ func Test_group_rule_operations(t *testing.T) {
 		Actions:    gra,
 		Conditions: grc,
 		Type:       "group_rule",
-		Name:       "Test group rule Updated",
+		Name:       "SDK_TEST group rule Updated",
 	}
 	newGroupRule, _, err := client.Group.UpdateGroupRule(ctx, groupRule.Id, *gr)
 	require.NoError(t, err, "Should not error when updating rule")