+++ tests/integration/request_test.go
@@ -38,7 +38,6 @@ func Test_private_key_request_contains_bearer_token(t *testing.T) {
 	require.NoError(t, err)
 
 	assert.Contains(t, request.Header.Get("Authorization"), "Bearer", "does not contain a bearer token for the request")
-
 }
 
 func Test_private_key_request_can_create_a_user(t *testing.T) {
@@ -51,11 +50,12 @@ func Test_private_key_request_can_create_a_user(t *testing.T) {
 	uc := &okta.UserCredentials{
 		Password: p,
 	}
+	email := randomEmail()
 	profile := okta.UserProfile{}
 	profile["firstName"] = "John"
 	profile["lastName"] = "Private_Key"
-	profile["email"] = "john-private-key@example.com"
-	profile["login"] = "john-private-key@example.com"
+	profile["email"] = email
+	profile["login"] = email
 	u := &okta.CreateUserRequest{
 		Credentials: uc,
 		Profile:     &profile,
@@ -67,7 +67,7 @@ func Test_private_key_request_can_create_a_user(t *testing.T) {
 	require.NoError(t, err, "Creating a new user should not error")
 	assert.NotEmpty(t, user.Id, "appears the user was not created")
 	tempProfile := *user.Profile
-	assert.Equal(t, "john-private-key@example.com", tempProfile["email"], "did not get the correct user")
+	assert.Equal(t, email, tempProfile["email"], "did not get the correct user")
 
 	// Deactivate the user → POST /api/v1/users/{{userId}}/lifecycle/deactivate
 	_, err = client.User.DeactivateUser(ctx, user.Id, nil)